+++ ProjectWeb.Domain/Models/WebProduct.cs
@@ -9,7 +9,7 @@ namespace ProjectWeb.Domain.Models
         public int WebProductID { get; set; }
         public string WebProductImage { get; set; }
         public string WebProductName { get; set; }
-        public double WebProductPrice { get; set; }
+        public long WebProductPrice { get; set; }
         public string WebProductDescription { get; set; }
         public string WebProductDeliverDate { get; set; }
     }
